Feat: Mejorados los potenciales errores

+++ Backend/src/main/java/org/pinguweb/backend/controller/AdminController.java
@@ -1,5 +1,6 @@
 package org.pinguweb.backend.controller;
 
+import org.pinguweb.backend.controller.common.ServerException;
 import org.pinguweb.backend.model.Admin;
 import org.pinguweb.backend.repository.AdminRepository;
 import org.springframework.http.ResponseEntity;
@@ -13,6 +14,8 @@ public class AdminController {
 
     @GetMapping("/admin/{id}")
     public ResponseEntity<Admin> getAdmin(@PathVariable String id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             return ResponseEntity.ok(repository.getReferenceById(id));
         }
@@ -23,11 +26,15 @@ public class AdminController {
 
     @PostMapping("/admin")
     public ResponseEntity<Admin> addAdmin(@RequestBody Admin admin) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         return ResponseEntity.ok(repository.save(admin));
     }
 
     @DeleteMapping("/admin/{id}")
     public ResponseEntity<Void>  deleteAdmin(@PathVariable String id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             repository.deleteById(id);
             return ResponseEntity.ok().build();
@@ -39,6 +46,8 @@ public class AdminController {
 
     @PutMapping("/admin")
     public ResponseEntity<Admin>  updateAdmin(@RequestBody Admin admin) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(admin.getDni())) {
             return ResponseEntity.ok(repository.save(admin));
         }
@@ -49,6 +58,8 @@ public class AdminController {
 
     @GetMapping("/admin/login/")
     public ResponseEntity<Void> login(@RequestParam String ID, @RequestParam String password) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(ID)) {
             Admin admin = repository.getReferenceById(ID);
             if (admin.getPassword().equals(password))
