+++ Backend/src/main/java/org/pinguweb/backend/controller/CatastropheController.java
@@ -1,9 +1,7 @@
 package org.pinguweb.backend.controller;
 
-import org.pinguweb.backend.model.Admin;
+import org.pinguweb.backend.controller.common.ServerException;
 import org.pinguweb.backend.model.Catastrophe;
-import org.pinguweb.backend.model.Need;
-import org.pinguweb.backend.repository.AdminRepository;
 import org.pinguweb.backend.repository.CatastropheRepository;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -18,12 +16,16 @@ public class CatastropheController {
 
     @GetMapping("/catastrophe")
     public ResponseEntity<List<Catastrophe>> getAll(){
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         List<Catastrophe> catastrophes = repository.findAll();
         return ResponseEntity.ok(catastrophes);
     }
 
     @GetMapping("/catastrophe/{id}")
     public ResponseEntity<Catastrophe> getCatastrophe(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             return ResponseEntity.ok(repository.getReferenceById(id));
         }
@@ -34,11 +36,15 @@ public class CatastropheController {
 
     @PostMapping("/catastrophe")
     public ResponseEntity<Catastrophe> addCatastrophe(@RequestBody Catastrophe catastrophe) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         return ResponseEntity.ok(repository.save(catastrophe));
     }
 
     @DeleteMapping("/catastrophe/{id}")
     public ResponseEntity<Void>  deleteCatastrophe(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             repository.deleteById(id);
             return ResponseEntity.ok().build();
@@ -50,6 +56,8 @@ public class CatastropheController {
 
     @PutMapping("/catastrophe")
     public ResponseEntity<Catastrophe>  updateCatastrophe(@RequestBody Catastrophe catastrophe) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(catastrophe.getID())) {
             return ResponseEntity.ok(repository.save(catastrophe));
         }
