+++ Backend/src/main/java/org/pinguweb/backend/controller/NeedController.java
@@ -1,5 +1,6 @@
 package org.pinguweb.backend.controller;
 
+import org.pinguweb.backend.controller.common.ServerException;
 import org.pinguweb.backend.model.Need;
 import org.pinguweb.backend.model.Need;
 import org.pinguweb.backend.repository.NeedRepository;
@@ -17,12 +18,16 @@ public class NeedController {
 
     @GetMapping("/need")
     public ResponseEntity<List<Need>> getAll(){
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         List<Need> needs = repository.findAll();
         return ResponseEntity.ok(needs);
     }
 
     @GetMapping("/need/{id}")
     public ResponseEntity<Need> getNeed(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             return ResponseEntity.ok(repository.getReferenceById(id));
         }
@@ -33,11 +38,15 @@ public class NeedController {
 
     @PostMapping("/need")
     public ResponseEntity<Need> addNeed(@RequestBody Need need) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         return ResponseEntity.ok(repository.save(need));
     }
 
     @DeleteMapping("/need/{id}")
     public ResponseEntity<Void>  deleteNeed(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             repository.deleteById(id);
             return ResponseEntity.ok().build();
@@ -49,6 +58,8 @@ public class NeedController {
 
     @PutMapping("/need")
     public ResponseEntity<Need>  updateNeed(@RequestBody Need need) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(need.getId())) {
             return ResponseEntity.ok(repository.save(need));
         }
