+++ Backend/src/main/java/org/pinguweb/backend/controller/TaskController.java
@@ -1,5 +1,6 @@
 package org.pinguweb.backend.controller;
 
+import org.pinguweb.backend.controller.common.ServerException;
 import org.pinguweb.backend.model.Task;
 import org.pinguweb.backend.repository.TaskRepository;
 import org.springframework.http.ResponseEntity;
@@ -15,12 +16,16 @@ public class TaskController {
 
     @GetMapping("/task")
     public ResponseEntity<List<Task>> getAll(){
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         List<Task> tasks = repository.findAll();
         return ResponseEntity.ok(tasks);
     }
 
     @GetMapping("/task/{id}")
     public ResponseEntity<Task> getTask(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             return ResponseEntity.ok(repository.getReferenceById(id));
         }
@@ -31,11 +36,15 @@ public class TaskController {
 
     @PostMapping("/task")
     public ResponseEntity<Task> addTask(@RequestBody Task task) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         return ResponseEntity.ok(repository.save(task));
     }
 
     @DeleteMapping("/task/{id}")
     public ResponseEntity<Void>  deleteTask(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             repository.deleteById(id);
             return ResponseEntity.ok().build();
@@ -47,6 +56,8 @@ public class TaskController {
 
     @PutMapping("/task")
     public ResponseEntity<Task>  updateTask(@RequestBody Task task) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(task.getId())) {
             return ResponseEntity.ok(repository.save(task));
         }
