+++ Backend/src/main/java/org/pinguweb/backend/controller/ZoneController.java
@@ -1,5 +1,6 @@
 package org.pinguweb.backend.controller;
 
+import org.pinguweb.backend.controller.common.ServerException;
 import org.pinguweb.backend.model.Zone;
 import org.pinguweb.backend.repository.ZoneRepository;
 import org.springframework.http.ResponseEntity;
@@ -15,12 +16,16 @@ public class ZoneController {
 
     @GetMapping("/zone")
     public ResponseEntity<List<Zone>> getAll(){
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         List<Zone> zones = repository.findAll();
         return ResponseEntity.ok(zones);
     }
 
     @GetMapping("/zone/{id}")
     public ResponseEntity<Zone> getZone(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             return ResponseEntity.ok(repository.getReferenceById(id));
         }
@@ -31,11 +36,15 @@ public class ZoneController {
 
     @PostMapping("/zone")
     public ResponseEntity<Zone> addZone(@RequestBody Zone zone) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         return ResponseEntity.ok(repository.save(zone));
     }
 
     @DeleteMapping("/zone/{id}")
     public ResponseEntity<Void>  deleteZone(@PathVariable Integer id) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(id)) {
             repository.deleteById(id);
             return ResponseEntity.ok().build();
@@ -47,6 +56,8 @@ public class ZoneController {
 
     @PutMapping("/zone")
     public ResponseEntity<Zone>  updateZone(@RequestBody Zone zone) {
+        if (!ServerException.isServerConnected(repository)){return ResponseEntity.internalServerError().build();}
+
         if (repository.existsById(zone.getId())) {
             return ResponseEntity.ok(repository.save(zone));
         }
